StartTLS should not just be pass-through.
Establishing a startTLS connection to the proxied service is something the
proxy operator needs to decide in advance.

If the proxy receives a startTLS request-- I haven't implemented that, so
the default behavior is for the proxy to return the "unavailable" (decimal 52)
response code as suggested by RFC4511 section 4.14.2.

The `handleStartTLSRequest()` method can be overridden to implement startTLS.

The proxy *can* currently be started using an SSL endpoint, so clients can use
LDAPS when connecting to it.  BINDs don't have to be in the clear.